Filter messages from blocked on muted accounts
At least for me, a lot of activity is importing posts from accounts I
have muted or blocked: usually those do automated posting at
high-traffic, and importing them is just useless work because we will
end up not seeing them anyway.

Load the mute and block lists at start time, then use them to filter
posts. Because of rate limiting, this slows down the time to first
import; but it evens out because we end up importing less useless posts.

Accounts are referenced by their URL, because it's the closest to a
universal identifier; using the account name would need re-formatting it
depending on server.

Finally, paging is not yet implemented, because I don't need it for now.